test: split slow test
