+++ tests/local-first-new-page.spec.ts
@@ -19,26 +19,4 @@ test.describe('local first new page', () => {
     const cell = page.getByRole('cell', { name: 'this is a new page' });
     expect(cell).not.toBeUndefined();
   });
-
-  test('click btn bew page and open in tab', async ({ page }) => {
-    await page.getByText('New Page').click();
-    await page.getByPlaceholder('Title').click();
-    await page.getByPlaceholder('Title').fill('this is a new page');
-    const newPageUrl = page.url();
-    const newPageId = page.url().split('/').reverse()[0];
-
-    await page.getByRole('link', { name: 'All pages' }).click();
-
-    await page
-      .getByTestId('more-actions-' + newPageId)
-      .getByRole('button')
-      .first()
-      .click();
-    const [newTabPage] = await Promise.all([
-      page.waitForEvent('popup'),
-      page.getByRole('button', { name: 'Open in new tab' }).click(),
-    ]);
-
-    expect(newTabPage.url()).toBe(newPageUrl);
-  });
 });
